+++ package.json
@@ -37,7 +37,6 @@
     "node": ">=0.12"
   },
   "dependencies": {
-    "vigour-js": "~0.3.8",
     "vigour-ua": "~0.0.3",
     "vigour-wrapper": "~2.2.1"
   },
