+++ test/browser/index.js
@@ -4,11 +4,11 @@ var tests = require('../tests')
 
 describe('Env', function () {
   describe('Mock Platform Tests', function () {
-    tests(require('./mockPlatform'), 'platform')
+    tests(require('../../lib/platform/mock'), 'platform')
   })
 
   describe('Mock Native Tests', function () {
-    var mockBridge = require('./mockNativeMethods')
+    var mockBridge = require('./mockBridge')
     var nativePlatform = require('../../lib/platform/native')
     tests(nativePlatform, mockBridge)
   })
